fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
ChemSpiPy==2.0.0
click==8.2.1
distro==1.9.0
fastapi==0.119.1
Flask==2.2.5
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==5.2.2
MarkupSafe==3.0.2
numpy==1.26.4
openai==1.52.2
ply==3.11
psycopg2-binary==2.9.9
PubChemPy==1.0.4
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-multipart==0.0.21
pygraphdb==2.0.1
PyJWT==2.10.1
Pyomo==6.8.0
pyparsing==3.2.5
python-dotenv==1.1.1
PyYAML==6.0.3
rdflib==7.2.1
rdkit-pypi
requests==2.32.3
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.32
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3